Rename 'ConfigurableX' to 'StandardX' for [Value]Graph, which better captures what these implementations represent.
Note that this work was already started in the past with the abstract StandardGraph tests.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=294879089
